feat: change download_rate_limit from double to uint64

+++ src/common.v2.rs
@@ -356,8 +356,8 @@ pub struct Download {
     #[prost(message, optional, tag = "12")]
     pub timeout: ::core::option::Option<::prost_wkt_types::Duration>,
     /// Download rate limit in bytes per second.
-    #[prost(double, optional, tag = "13")]
-    pub download_rate_limit: ::core::option::Option<f64>,
+    #[prost(uint64, optional, tag = "13")]
+    pub download_rate_limit: ::core::option::Option<u64>,
     /// NeedBackToSource needs downloaded from source.
     #[prost(bool, tag = "14")]
     pub need_back_to_source: bool,
